feat: add `useSavePendingArtworkListsChanges` hook
